third_party/blink/web_tests/platform/generic/fast/css/invalidation/enabled-pseudo-expected.txt

Use descendant invalidation set for :enabled pseudo class.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getComputedStyle(button, '').backgroundColor is transparent
PASS internals.updateStyleAndReturnAffectedElementCount() is 1
PASS getComputedStyle(button, '').backgroundColor is green
PASS successfullyParsed is true

TEST COMPLETE
